requirements.txt
python-telegram-bot==13.1
bs4==0.0.1
requests==2.25.1